Fix a non-deterministic build bug
This was probably one of the most painful debugging sessions I've done
so far.

When bootstrapping the OCaml compiler with the JIT enabled, builds would
randomly fail due to an out-of-memory condition (fatal error).

It wouldn't always be on the same file and the amount of time it took to
trigger was random but it was nearly almost every build at least onece.

Getting a debugger was way too hard - all of this was print debugging.

I tracked this down to malloc failing. The trace upwards was:

memory.c:

1. caml_stat_alloc_noexc fails because the requested size is too big
2. caml_stat_alloc_aligned_noexc was given too big a size
3. caml_alloc_for_heap was given too big a request
4. expand_heap was given too big a request
5. caml_alloc_shr_aux was given too big a wosize and the heap was full
6. caml_alloc_shr_for_minor_gc too big a wosize

Then we go into:

minor_gc.c:

7. caml_oldify_one thought something was a valid pointer to something in
  the minor heap but the header pointer had a size that was way too big to
  be something realistic

At this point I was stuck for ages. I turned on verbose tracing with
OCAML_RUNPARAM and eventually worked out it was happening during a
specific phase of the GC cycle. Going top-down:

1. A GC is triggered calling caml_empty_minor_heap
2. Which calls caml_final_update_minor_roots

roots_byt.c:

3. which calls caml_oldify_one on everything in the stack

So the preconditions were:

- During a GC cycle when the GC is out of memory
- There's a value on the stack which is a valid pointer into the minor
  heap
- This value's header has a way too large wosize

This is something I could check for - I added some code to hook into
caml_final_update_minor_roots and detect this as a problem and print the
stack if it sees it.

I then saw the position of the problem was always sp[3].

This made me remember Setup_for_gc - before every call that could
trigger a garbage collection we make sure to push the current accu and
env values onto the stack so that the GC can fix them.

The bytecode interpreter also pushes the PC for the purpose of
backtraces (if there was an exception during the finalisation of
something during the GC for example). So it drops sp by 3 and pushes
those three values.

I thought that I was maybe not initialising sp[3] when I do this because
I don't bother saving and restoring the pc because I don't care about
backtraces. Sure enough this is what had happened and the small diff in
this commit  was the fix.